Removed extra logic when loading the filter file
Removed extra logic that checked stuff like the types of each element in the filter file on boot. The function now was renamed to check_filter_file and it just makes sure the file exists and if it doesn't, generates the file. 2nd aliases for the condition add, edit, and remove commands were also added.